modulare cart and fixed new path issue
